Revert "In progress"
This reverts commit 99df3b0b9f8ae5e44323d98cad8bbbca157ea8d7.

+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/OID.java
@@ -59,8 +59,8 @@ public abstract class OID extends TSBeanBase implements Authorizable{
 	}
 
 	@Override
-	public long getPk() {
-		return (int) pk;
+	public int getPk() {
+		return pk;
 	}
 
 	public void setPk(long pk) {
